os2: kill an execvp()ed child correctly when a parent exits
On OS/2, execvp() does not overlay a parent process. Instead, a parent
waits for a child to quit. As a result, kill() terminates a fork()ed
parent not a execvp()ed child with a pid returned by fork(), and the
child remains as a zombie process. Because of this, a shell cannot be
foreground until an user terminates a child by pressing Ctrl-C or
Ctrl-Break.

Fixed #4.